Storybook: Try to integrate the playground into Storybook (#18191)
* Storybook: Try to integrate the playground into Storybook

* Load Noto font in storybook via preview-head.html

* Add scss support for Storybook!

* Remove unused parcel-bundler

* Update mentions of the playground with Storybook

* Fix issues found during self-review

* Add back playground:* aliasing storybook:* for the time being

* Add help info for old commands

* Ensure that source for the playground can be presented in Storybook

* Move and update the README document added for Storybook

* Resolve all issues raised in the review
