+++ .travis.yml
@@ -203,8 +203,7 @@ jobs:
       install:
         - npm ci
       before_deploy:
-        - npm run playground:build -- --public-url '/gutenberg'
-        - npm run design-system:build
+        - npm run storybook:build
       deploy:
         provider: pages
         skip_cleanup: true
