+++ package.json
@@ -108,6 +108,7 @@
 		"concurrently": "3.5.0",
 		"copy-webpack-plugin": "4.5.2",
 		"cross-env": "3.2.4",
+		"css-loader": "3.2.0",
 		"cssnano": "4.1.10",
 		"deep-freeze": "0.0.1",
 		"enzyme": "3.9.0",
@@ -132,7 +133,6 @@
 		"mkdirp": "0.5.1",
 		"node-sass": "4.12.0",
 		"node-watch": "0.6.0",
-		"parcel-bundler": "1.12.4",
 		"postcss": "7.0.13",
 		"progress": "2.0.3",
 		"react": "16.9.0",
@@ -149,6 +149,7 @@
 		"simple-git": "1.113.0",
 		"source-map-loader": "0.2.4",
 		"sprintf-js": "1.1.1",
+		"style-loader": "1.0.0",
 		"stylelint-config-wordpress": "13.1.0",
 		"typescript": "3.5.3",
 		"uuid": "3.3.2",
@@ -198,10 +199,14 @@
 		"test-unit-php-multisite": "cross-env WP_MULTISITE=1 wp-scripts env test-php",
 		"test-unit:native": "cd test/native/ && cross-env NODE_ENV=test jest --config ./jest.config.js",
 		"test-unit:native:debug": "cd test/native/ && node --inspect ../../node_modules/.bin/jest --runInBand --config ./jest.config.js",
-		"design-system:build": "npm run build:packages && build-storybook -c ./packages/components/storybook -o ./playground/dist/design-system/components",
-		"design-system:dev": "npm run build:packages && concurrently \"npm run dev:packages\" \"start-storybook -c ./packages/components/storybook\"",
-		"playground:build": "npm run build:packages && parcel build playground/src/index.html -d playground/dist",
-		"playground:dev": "concurrently \"npm run dev:packages\" \"parcel playground/src/index.html -d playground/dist\"",
+		"prestorybook:build": "npm run build:packages",
+		"storybook:build": "build-storybook -c ./storybook -o ./playground/dist",
+		"prestorybook:dev": "npm run build:packages",
+		"storybook:dev": "concurrently \"npm run dev:packages\" \"start-storybook -c ./storybook\"",
+		"design-system:build": "echo \"Please use storybook:build instead.\"",
+		"design-system:dev": "echo \"Please use storybook:dev instead.\"",
+		"playground:build": "npm run storybook:build",
+		"playground:dev": "echo \"Please use storybook:dev instead.\"",
 		"preenv": "npm run check-engines",
 		"env": "wp-scripts env"
 	},
