+++ packages/components/src/modal/stories/index.js
@@ -15,7 +15,7 @@ import Modal from '../';
  */
 import { useState } from '@wordpress/element';
 
-export default { title: 'Modal', component: Modal };
+export default { title: 'Components|Modal', component: Modal };
 
 const ModalExample = ( props ) => {
 	const [ isOpen, setOpen ] = useState( false );
