+++ packages/components/src/scroll-lock/stories/index.js
@@ -9,7 +9,7 @@ import { useState } from '@wordpress/element';
 import Button from '../../button';
 import ScrollLock from '../';
 
-export default { title: 'ScrollLock', component: ScrollLock };
+export default { title: 'Components|ScrollLock', component: ScrollLock };
 
 const Example = () => {
 	const [ isScrollLocked, setScrollLocked ] = useState( false );
