+++ packages/components/src/snackbar/stories/index.js
@@ -8,7 +8,7 @@ import { text } from '@storybook/addon-knobs';
  */
 import Snackbar from '../';
 
-export default { title: 'Snackbar', component: Snackbar };
+export default { title: 'Components|Snackbar', component: Snackbar };
 
 export const _default = () => {
 	const content = text( 'Content', 'Use Snackbars to communicate low priority, non-interruptive messages to the user.' );
