--- packages/components/storybook/.babelrc
@@ -1,3 +1,0 @@
-{
-	"presets": [ "@wordpress/babel-preset-default" ]
-}
