--- packages/components/storybook/README.md
@@ -1,11 +1,0 @@
-
-# Storybook 
-
-Storybook is an open-source tool that provides a sandbox to develop and visualize components in isolation. See the [Storybookjs site](https://storybook.js.org/) for more information about the tool. 
-
-The Gutenberg project uses Storybook to view and work with the `@wordpress/components` package.
-
-View online at: https://wordpress.github.io/gutenberg/design-system/components
-
-Run locally in your development environment running: `npm run design-system:dev` from the top-level Gutenberg directory.
-
