--- packages/components/storybook/webpack.config.js
@@ -1,9 +1,0 @@
-module.exports = ( { config } ) => {
-	config.module.rules.push( {
-		test: /\/stories\/.+\.js$/,
-		loaders: [ require.resolve( '@storybook/source-loader' ) ],
-		enforce: 'pre',
-	} );
-
-	return config;
-};
