--- playground/.babelrc
@@ -1,9 +1,0 @@
-{
-	"presets": ["@wordpress/babel-preset-default"],
-	"plugins": [
-		[ "@babel/plugin-transform-react-jsx", {
-			"pragma": "createElement"
-		} ],
-		"babel-plugin-inline-json-import"
-	]
-}
