--- playground/src/index.html
@@ -1,13 +1,0 @@
-<!doctype html>
-<html lang="en">
-<head>
-	<meta charset="utf-8">
-	<title>The Gutenberg Playground</title>
-	<link href="https://fonts.googleapis.com/css?family=Noto+Serif:400,700" rel="stylesheet">
-</head>
-
-<body>
-	<div id="app"></div>
-	<script src="./index.js"></script>
-</body>
-</html>
