+++ storybook/.babelrc
@@ -1,0 +1,6 @@
+{
+	"presets": [ "@wordpress/babel-preset-default" ],
+	"plugins": [
+		"babel-plugin-inline-json-import"
+	]
+}
