+++ storybook/config.js
@@ -1,0 +1,23 @@
+/**
+ * External dependencies
+ */
+import { addDecorator, configure } from '@storybook/react';
+import { withA11y } from '@storybook/addon-a11y';
+import { withKnobs } from '@storybook/addon-knobs';
+
+/**
+ * WordPress dependencies
+ */
+/* eslint-disable no-restricted-syntax */
+import '@wordpress/components/build-style/style.css';
+/* eslint-enable no-restricted-syntax */
+
+addDecorator( withA11y );
+addDecorator( withKnobs );
+configure(
+	[
+		require.context( './stories/', true, /\/.+\.(js|mdx)$/ ),
+		require.context( '../packages/components/src/', true, /\/stories\/.+\.js$/ ),
+	],
+	module
+);
