+++ storybook/preview-head.html
@@ -1,0 +1,1 @@
+<link href="https://fonts.googleapis.com/css?family=Noto+Serif:400,700" rel="stylesheet">
