+++ storybook/stories/docs/introduction.story.mdx
@@ -1,6 +1,6 @@
 import { Meta } from '@storybook/addon-docs/blocks';
 
-<Meta title="Introduction" />
+<Meta title="Docs|Introduction" />
 
 # Introduction
